+++ packages/grafana-runtime/package.json
@@ -18,6 +18,7 @@
   "license": "Apache-2.0",
   "dependencies": {},
   "devDependencies": {
+    "@types/systemjs": "^0.20.6",
     "awesome-typescript-loader": "^5.2.1",
     "lodash": "^4.17.10",
     "pretty-format": "^24.5.0",
