+++ public/app/core/config.ts
@@ -1,79 +1,5 @@
-import _ from 'lodash';
-import { GrafanaTheme, getTheme, GrafanaThemeType, PanelPluginMeta, DataSourceInstanceSettings } from '@grafana/ui';
+import { config, GrafanaBootConfig } from '@grafana/runtime';
 
-export interface BuildInfo {
-  version: string;
-  commit: string;
-  isEnterprise: boolean;
-  env: string;
-  latestVersion: string;
-  hasUpdate: boolean;
-}
-
-export class Settings {
-  datasources: { [str: string]: DataSourceInstanceSettings };
-  panels: { [key: string]: PanelPluginMeta };
-  appSubUrl: string;
-  windowTitlePrefix: string;
-  buildInfo: BuildInfo;
-  newPanelTitle: string;
-  bootData: any;
-  externalUserMngLinkUrl: string;
-  externalUserMngLinkName: string;
-  externalUserMngInfo: string;
-  allowOrgCreate: boolean;
-  disableLoginForm: boolean;
-  defaultDatasource: string;
-  alertingEnabled: boolean;
-  alertingErrorOrTimeout: string;
-  alertingNoDataOrNullValues: string;
-  authProxyEnabled: boolean;
-  exploreEnabled: boolean;
-  ldapEnabled: boolean;
-  oauth: any;
-  disableUserSignUp: boolean;
-  loginHint: any;
-  passwordHint: any;
-  loginError: any;
-  viewersCanEdit: boolean;
-  editorsCanAdmin: boolean;
-  disableSanitizeHtml: boolean;
-  theme: GrafanaTheme;
-  pluginsToPreload: string[];
-
-  constructor(options: Settings) {
-    this.theme = options.bootData.user.lightTheme ? getTheme(GrafanaThemeType.Light) : getTheme(GrafanaThemeType.Dark);
-
-    const defaults = {
-      datasources: {},
-      windowTitlePrefix: 'Grafana - ',
-      panels: {},
-      newPanelTitle: 'Panel Title',
-      playlist_timespan: '1m',
-      unsaved_changes_warning: true,
-      appSubUrl: '',
-      buildInfo: {
-        version: 'v1.0',
-        commit: '1',
-        env: 'production',
-        isEnterprise: false,
-      },
-      viewersCanEdit: false,
-      editorsCanAdmin: false,
-      disableSanitizeHtml: false,
-    };
-
-    _.extend(this, defaults, options);
-  }
-}
-
-const bootData = (window as any).grafanaBootData || {
-  settings: {},
-  user: {},
-};
-
-const options = bootData.settings;
-options.bootData = bootData;
-
-export const config = new Settings(options);
+// Legacy binding paths
+export { config, GrafanaBootConfig as Settings };
 export default config;
