+++ public/app/core/utils/ConfigProvider.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import config, { Settings } from 'app/core/config';
+import { config, GrafanaBootConfig } from '@grafana/runtime';
 import { GrafanaThemeType, ThemeContext, getTheme } from '@grafana/ui';
 
-export const ConfigContext = React.createContext<Settings>(config);
+export const ConfigContext = React.createContext<GrafanaBootConfig>(config);
 export const ConfigConsumer = ConfigContext.Consumer;
 
 export const provideConfig = (component: React.ComponentType<any>) => {
