+++ public/app/plugins/sdk.ts
@@ -2,6 +2,6 @@ import { PanelCtrl } from 'app/features/panel/panel_ctrl';
 import { MetricsPanelCtrl } from 'app/features/panel/metrics_panel_ctrl';
 import { QueryCtrl } from 'app/features/panel/query_ctrl';
 import { alertTab } from 'app/features/alerting/AlertTabCtrl';
-import { loadPluginCss } from 'app/features/plugins/plugin_loader';
+import { loadPluginCss } from '@grafana/runtime';
 
 export { PanelCtrl, MetricsPanelCtrl, QueryCtrl, alertTab, loadPluginCss };
